Tried to use BehaviorSubject, currently not working.
Using a EventEmitter to pass data from one component to the other does
not seem to be the ideal way. Tried to use a BehaviorSubject insted,
which is not working as expected. During construciton, the first tab is
set correctly, however, pushing the buttons does not trigger the
Observer functionality...